refactor: BubbleContent, EventOverlay 분리
