+++ src/components/game-room/boardMap/GameBoard.tsx
@@ -1,18 +1,13 @@
 import { useEffect, useRef, useState } from 'react';
 
-import Image from 'next/image';
-
-import { cn } from '@/lib/utils';
 import { useSoundEventStore } from '@/stores/useSoundEventStore';
-import {
-  EventType,
-  useGameBubbleStore,
-} from '@/stores/websocket/useGameBubbleStore';
+import { useGameBubbleStore } from '@/stores/websocket/useGameBubbleStore';
 import { useGameDiceStore } from '@/stores/websocket/useGameDiceStore';
 import { useParticipantInfoStore } from '@/stores/websocket/useGameParticipantStore';
 import { useScoreStore } from '@/stores/websocket/useScoreStore';
 
-import EventCard from './EventCard';
+import BubbleContent from './BubbleContent';
+import EventOverlay from './EventOverlay';
 
 interface FootholderPosition {
   xRatio: number;
@@ -32,7 +27,6 @@ interface UserCharacter {
   moveStartTime: number;
 }
 
-// 테두리에 50px 패딩을 적용하고 좌우 대칭을 맞춘 footholderRatios
 const footholderRatios: FootholderPosition[] = [
   { xRatio: 0.1, yRatio: 0.88, size: 2 }, // 0번 위치
   { xRatio: 0.23, yRatio: 0.86, size: 1.5 }, // 1번 위치
@@ -70,111 +64,6 @@ const bubbleRightMap: Record<number, boolean> = {
   15: true,
   16: true,
   17: true,
-};
-
-const EventOverlay = ({ eventType }: { eventType: EventType }) => {
-  return (
-    <div className='event-overlay'>
-      <div
-        className={cn('flip-card-container w-full', {
-          'animate-scale-in animate-flip': eventType !== 'MARK',
-          'animate-scale-in': eventType === 'MARK',
-        })}
-        style={{ transformStyle: 'preserve-3d' }}
-      >
-        <div
-          className='flip-card-front'
-          style={{ backfaceVisibility: 'hidden' }}
-        >
-          <EventCard eventType={'MARK'} />
-        </div>
-
-        <div
-          className='flip-card-back'
-          style={{ backfaceVisibility: 'hidden' }}
-        >
-          <EventCard eventType={eventType} />
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const BubbleContent = ({ isActive }: { isActive: boolean }) => {
-  const [currentImage, setCurrentImage] = useState(1);
-  const [isAnimating, setIsAnimating] = useState(false);
-  const { diceTotalmovement } = useGameDiceStore();
-  const animationRef = useRef<NodeJS.Timeout | null>(null);
-  const { setSoundEvent } = useSoundEventStore();
-
-  const imagePaths = [
-    '/eventemoji/move_1.png',
-    '/eventemoji/move_2.png',
-    '/eventemoji/move_3.png',
-  ];
-
-  useEffect(() => {
-    if (isActive && !isAnimating) {
-      setIsAnimating(true);
-      setSoundEvent('ROULETTE_123');
-
-      let count = 0;
-      const startAnimation = () => {
-        animationRef.current = setInterval(() => {
-          count++;
-          setCurrentImage(prev => (prev % 3) + 1);
-
-          if (count >= 6) {
-            if (animationRef.current) {
-              clearInterval(animationRef.current);
-            }
-            if (diceTotalmovement) {
-              setCurrentImage(diceTotalmovement);
-              setSoundEvent('ROULETTE_123_RESULT');
-            }
-            setIsAnimating(false);
-          }
-        }, 167);
-      };
-
-      startAnimation();
-    }
-
-    return () => {
-      if (animationRef.current) {
-        clearInterval(animationRef.current);
-      }
-    };
-  }, [isActive, diceTotalmovement]);
-
-  useEffect(() => {
-    if (!isActive) {
-      setIsAnimating(false);
-      if (animationRef.current) {
-        clearInterval(animationRef.current);
-      }
-    }
-  }, [isActive]);
-
-  if (!diceTotalmovement) return null;
-
-  return (
-    <div className='relative flex h-full w-full items-center justify-center'>
-      {imagePaths.map((path, index) => (
-        <Image
-          key={index}
-          className={cn(
-            currentImage !== index + 1 ? 'opacity-0' : 'opacity-100',
-            `absolute top-1/4 left-1/4 -mt-2 h-1/2 w-1/2 object-contain transition-opacity duration-100 ease-in-out`,
-          )}
-          src={path}
-          alt={`Move ${index + 1}`}
-          width={200}
-          height={200}
-        />
-      ))}
-    </div>
-  );
 };
 
 const GameBoard = () => {
